LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-2-expected.txt


FAIL parse error has higher priority than instantiation error assert_equals: expected 4 but got 3
